add a keep option in phrase command

+++ symfony/src/Command/PhraseCommand.php
@@ -45,7 +45,8 @@ class PhraseCommand extends Command
             ->setDescription('Synchronize translations with PhraseApp')
             ->addOption('sleep', null, InputOption::VALUE_OPTIONAL, 'Use this option for large operations to prevent hitting rate limits', 0)
             ->addOption('delete', null, InputOption::VALUE_NONE, 'Add this option to automatically delete translations that are on Phrase but no more on the app')
-            ->addOption('create', null, InputOption::VALUE_NONE, 'Add this option to automatically create translations that are on the app but not yet on Phrase');
+            ->addOption('create', null, InputOption::VALUE_NONE, 'Add this option to automatically create translations that are on the app but not yet on Phrase')
+            ->addOption('keep', null, InputOption::VALUE_NONE, 'Add this option to prevent updating local files.');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -121,12 +122,14 @@ class PhraseCommand extends Command
         }
 
         // Dumping files
-        foreach ($remoteTranslations as $file => $keys) {
-            $oldContent = is_file($file) ? file_get_contents($file) : null;
-            $newContent = Yaml::dump($this->getDeflattedTranslationsFromArray($keys), 64, 2);
-            if ($oldContent !== $newContent) {
-                file_put_contents($file, $newContent);
-                $output->writeln(sprintf('Translations updated: %s', $file));
+        if (!$input->getOption('keep')) {
+            foreach ($remoteTranslations as $file => $keys) {
+                $oldContent = is_file($file) ? file_get_contents($file) : null;
+                $newContent = Yaml::dump($this->getDeflattedTranslationsFromArray($keys), 64, 2);
+                if ($oldContent !== $newContent) {
+                    file_put_contents($file, $newContent);
+                    $output->writeln(sprintf('Translations updated: %s', $file));
+                }
             }
         }
     }
